Fix selective extractor stats
As the selective extractor is not configured via rules there are also
no statistics of matched rules in the processor stats. Previously
this was not considered. Now to account for this the processor stats
module ignores the 'matches_per_idx' and 'times_per_idx' stats for
the selective extractor processor.