Improve the domain label extraction processor
The unrecognized domain label error has been removed. If an ip
address is present the processor does nothing. If neither an
ip nor a domain is present a tag "unrecognized_domain" is added
to the event. Furthermore the domain label extraction processor
can now be configured with a tld suffix list.