Improve list comparison performance
The list comparison processor did not make use of the performance of
sets or dictionaries as it was still iterating over every element.
This new implementation uses a slightly different datastructure to
speed up the matching of elements and lists. It now only iterates
over the number of lists given in a rule instead of each element
of all lists. This significantly speeds up the worst case scenario.