Add more informative domain label extractor tags
The previous tags only included one error tag when no domain was
recognized. With this commit tags will also be added when an IP
address was found in the target field. Now the added tags also
include the target field name such that it is possible to
distinguish between different fields and tags.