Fix init of stats in processor_stats.py
Two statistics were only initialized when rules where added to the
processor. This leads to an error if a processor with zero rules is
currently active in the pipeline. Even though it is uncommon to use
processors without a rule it can be still helpful while designing a
pipeline for which the rules will be added in a later step.