Fix typo in processor_stats.py (#78)
* Fix init of stats in processor_stats.py

Two statistics were only initialized when rules where added to the
processor. This leads to an error if a processor with zero rules is
currently active in the pipeline. Even though it is uncommon to use
processors without a rule it can be still helpful while designing a
pipeline for which the rules will be added in a later step.

* Update test to account for new init

The tests had to be changed such that they also check if the stats
matches_per_idx and times_per_idx are initialized properly.

* Add camel_to_snake converter

* Add SnakeType to BaseProcessor

* Revert first fix and implement actual fix

The init was not the problem, there was a typo in a processor type
name. This commit introduces a list of processor names as rule based
exclusions such that the risk of more typos is reduced.

* Fix linting

* Fix black formatting

* Fix typo

* Refactor test

Use variables instead of magic numbers
Add assertion message to assertions
Add three more test cases to improve the exclusion test

Co-authored-by: ekneg54 <ekneg54@pm.me>