Refactor prometheus metrics
To improve the visualizations in grafana a few changes have been
made. First and foremost stale metric database files will be removed
now when a pipeline fails. Furthermore, the metrics are prefixed
with 'logprep_' and the label is now called 'component'.
Additionally, the status logger is not stored in a list anymore
but in a namedtuple, which makes the unwrapping easier. Last but
not least a small fix in the documentation was done, variables
renamed and some linting issues fixed.