Move hmac config options
The hmac config options are now part of the preprocessor. Backwards
capability is maintained for now by allowing the hmac options in two
positions. The old position under 'consumer' will lead to a
deprecation warning. If the old position is used it will be
automatically copied to the new location under preprocessor.