Add logprep and configuration version to event (#149)
* Add logprep and preprocessing version to event

This enables the pipeline to automatically add version information
to all events independent of the pipeline configuration.

* Update CHANGELOG.md

* Move hmac config options

The hmac config options are now part of the preprocessor. Backwards
capability is maintained for now by allowing the hmac options in two
positions. The old position under 'consumer' will lead to a
deprecation warning. If the old position is used it will be
automatically copied to the new location under preprocessor.

* Add deprecation comments

Add a new section in the CHANGELOG.md to communicate upcoming breaking
changes that do not have to be in the next release.

* Fix deprecation message in test

* Implement review comments

* add log and sql_db_table.json to .gitignore

* Remove check for unknown options

* Add designated test for update_default_configuration

* delete redundant test for confluent kafka