fix broken gracefull shutdown (#568)
* fix gracefull timout

had to increase the `timeout_graceful_shutdown` internal config parameter.
without this the server drops an error message on every shutdown.

hat to check for attribute `http_server` before shutdown, because the shutdown
proceeds in every process, but only in the first process this attribute is defined

* only start multiprocess logging in main process

had to ensure to only start multiprocess logging in main
process by checking for the current process name. This is needed
because of forking for pipeline creation is complaining for a deamonic process
having a child (the log process).

* register signals in cli function

the cli() function is the main entrypoint for logprep by installing
a script called "logprep" via setuptools. so nothing in the root of the module
run_logprep is executed